convert servlet 3.0 tests to java
