+++ instrumentation/servlet/servlet-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/jetty/JettyServlet3AsyncTest.java
@@ -8,9 +8,19 @@ package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.jetty;
 import io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.TestServlet3;
 import javax.servlet.Servlet;
 
-public class JettyServlet3TestSyncTest extends JettyServlet3Test {
+public class JettyServlet3AsyncTest extends JettyServlet3Test {
   @Override
   public Class<? extends Servlet> servlet() {
-    return TestServlet3.Sync.class;
+    return TestServlet3.Async.class;
+  }
+
+  @Override
+  public boolean errorEndpointUsesSendError() {
+    return false;
+  }
+
+  @Override
+  public boolean isAsyncTest() {
+    return true;
   }
 }
