+++ instrumentation/servlet/servlet-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/jetty/JettyServlet3FakeAsyncTest.java
@@ -8,7 +8,7 @@ package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.jetty;
 import io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.TestServlet3;
 import javax.servlet.Servlet;
 
-public class JettyServlet3TestFakeAsyncTest extends JettyServlet3Test {
+public class JettyServlet3FakeAsyncTest extends JettyServlet3Test {
   @Override
   public Class<? extends Servlet> servlet() {
     return TestServlet3.FakeAsync.class;
