+++ instrumentation/servlet/servlet-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/jetty/JettyServlet3SyncTest.java
@@ -1,0 +1,16 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.jetty;
+
+import io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.TestServlet3;
+import javax.servlet.Servlet;
+
+public class JettyServlet3SyncTest extends JettyServlet3Test {
+  @Override
+  public Class<? extends Servlet> servlet() {
+    return TestServlet3.Sync.class;
+  }
+}
