+++ instrumentation/servlet/servlet-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/jetty/JettyServletHandlerTest.java
@@ -56,7 +56,7 @@ public class JettyServletHandlerTest extends AbstractServlet3Test<Server, Servle
       SpanDataAssert span, SpanData parentSpan, String method, ServerEndpoint endpoint) {
 
     if (JettyServlet3Test.IS_BEFORE_94 && endpoint.equals(EXCEPTION)) {
-      span.satisfies(it -> assertThat(it.getName()).matches("\\.sendError$"))
+      span.satisfies(it -> assertThat(it.getName()).matches(".*\\.sendError"))
           .hasKind(SpanKind.INTERNAL)
           .hasParent(parentSpan);
     }
