+++ instrumentation/servlet/servlet-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/tomcat/TomcatServlet3AsyncTest.java
@@ -3,12 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.jetty;
+package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat;
 
-import io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.TestServlet3;
 import javax.servlet.Servlet;
 
-public class JettyServlet3TestAsyncTest extends JettyServlet3Test {
+public class TomcatServlet3AsyncTest extends TomcatServlet3Test {
   @Override
   public Class<? extends Servlet> servlet() {
     return TestServlet3.Async.class;
@@ -17,10 +16,5 @@ public class JettyServlet3TestAsyncTest extends JettyServlet3Test {
   @Override
   public boolean errorEndpointUsesSendError() {
     return false;
-  }
-
-  @Override
-  public boolean isAsyncTest() {
-    return true;
   }
 }
