--- instrumentation/servlet/servlet-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/tomcat/TomcatServlet3TestFakeAsyncTest.java
@@ -1,15 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat;
-
-import javax.servlet.Servlet;
-
-public class TomcatServlet3TestFakeAsyncTest extends TomcatServlet3Test {
-  @Override
-  public Class<? extends Servlet> servlet() {
-    return TestServlet3.FakeAsync.class;
-  }
-}
