+++ instrumentation/servlet/servlet-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/tomcat/dispatch/TomcatServlet3DispatchAsyncTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.jetty.dispatch;
+package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.dispatch;
 
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.AUTH_REQUIRED;
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.CAPTURE_HEADERS;
@@ -17,28 +17,18 @@ import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint
 
 import io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.TestServlet3;
 import javax.servlet.Servlet;
-import org.eclipse.jetty.servlet.ServletContextHandler;
+import org.apache.catalina.Context;
 
-public class JettyServlet3TestDispatchAsyncTest extends JettyDispatchTest {
+public class TomcatServlet3DispatchAsyncTest extends TomcatDispatchTest {
   @Override
   public Class<? extends Servlet> servlet() {
     return TestServlet3.Async.class;
   }
 
   @Override
-  public boolean isAsyncTest() {
-    return true;
-  }
-
-  @Override
-  protected void setupServlets(ServletContextHandler context) {
+  protected void setupServlets(Context context) {
     super.setupServlets(context);
-    addServlet(
-        context, "/dispatch" + HTML_PRINT_WRITER.getPath(), TestServlet3.DispatchAsync.class);
-    addServlet(
-        context,
-        "/dispatch" + HTML_SERVLET_OUTPUT_STREAM.getPath(),
-        TestServlet3.DispatchAsync.class);
+
     addServlet(context, "/dispatch" + SUCCESS.getPath(), TestServlet3.DispatchAsync.class);
     addServlet(context, "/dispatch" + QUERY_PARAM.getPath(), TestServlet3.DispatchAsync.class);
     addServlet(context, "/dispatch" + ERROR.getPath(), TestServlet3.DispatchAsync.class);
@@ -49,6 +39,12 @@ public class JettyServlet3TestDispatchAsyncTest extends JettyDispatchTest {
     addServlet(
         context, "/dispatch" + CAPTURE_PARAMETERS.getPath(), TestServlet3.DispatchAsync.class);
     addServlet(context, "/dispatch" + INDEXED_CHILD.getPath(), TestServlet3.DispatchAsync.class);
+    addServlet(
+        context, "/dispatch" + HTML_PRINT_WRITER.getPath(), TestServlet3.DispatchAsync.class);
+    addServlet(
+        context,
+        "/dispatch" + HTML_SERVLET_OUTPUT_STREAM.getPath(),
+        TestServlet3.DispatchAsync.class);
     addServlet(context, "/dispatch/recursive", TestServlet3.DispatchRecursive.class);
   }
