+++ instrumentation/servlet/servlet-3.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/tomcat/dispatch/TomcatServlet3ForwardTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.jetty.dispatch;
+package io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.dispatch;
 
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.AUTH_REQUIRED;
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.CAPTURE_HEADERS;
@@ -15,28 +15,29 @@ import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.REDIRECT;
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.SUCCESS;
 
+import io.opentelemetry.instrumentation.testing.junit.http.HttpServerTestOptions;
 import io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.RequestDispatcherServlet;
 import io.opentelemetry.javaagent.instrumentation.servlet.v3_0.tomcat.TestServlet3;
 import javax.servlet.Servlet;
-import org.eclipse.jetty.servlet.ServletContextHandler;
+import org.apache.catalina.Context;
 
-public class JettyServlet3TestForwardTest extends JettyDispatchTest {
+public class TomcatServlet3ForwardTest extends TomcatDispatchTest {
   @Override
   public Class<? extends Servlet> servlet() {
     return TestServlet3.Sync.class; // dispatch to sync servlet
   }
 
   @Override
-  protected void setupServlets(ServletContextHandler context) {
+  protected void configure(HttpServerTestOptions options) {
+    super.configure(options);
+    options.setTestNotFound(false);
+  }
+
+  @Override
+  protected void setupServlets(Context context) {
     super.setupServlets(context);
 
     addServlet(context, "/dispatch" + SUCCESS.getPath(), RequestDispatcherServlet.Forward.class);
-    addServlet(
-        context, "/dispatch" + HTML_PRINT_WRITER.getPath(), RequestDispatcherServlet.Forward.class);
-    addServlet(
-        context,
-        "/dispatch" + HTML_SERVLET_OUTPUT_STREAM.getPath(),
-        RequestDispatcherServlet.Forward.class);
     addServlet(
         context, "/dispatch" + QUERY_PARAM.getPath(), RequestDispatcherServlet.Forward.class);
     addServlet(context, "/dispatch" + REDIRECT.getPath(), RequestDispatcherServlet.Forward.class);
@@ -52,5 +53,11 @@ public class JettyServlet3TestForwardTest extends JettyDispatchTest {
         RequestDispatcherServlet.Forward.class);
     addServlet(
         context, "/dispatch" + INDEXED_CHILD.getPath(), RequestDispatcherServlet.Forward.class);
+    addServlet(
+        context, "/dispatch" + HTML_PRINT_WRITER.getPath(), RequestDispatcherServlet.Forward.class);
+    addServlet(
+        context,
+        "/dispatch" + HTML_SERVLET_OUTPUT_STREAM.getPath(),
+        RequestDispatcherServlet.Forward.class);
   }
 }
